schedule-mcp-server/pyproject.toml
[project]
name = "schedule"
version = "0.1.0"
description = "Schedule MCP Server - 定时任务和大屏推送服务"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.27.0",
    "httpx>=0.28.1",
    "croniter>=2.0.0",
    "apscheduler>=3.10.0",
    "arrow>=1.3.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",           # 新增：SQLAlchemy ORM
    "aiosqlite>=0.19.0",           # 新增：异步 SQLite 支持
]

[project.scripts]
schedule-mcp = "schedule.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling]
package-dir = "src"
packages = ["schedule"]